jira_agent/requirements.txt
jira==3.5.1
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.34 